requirements.txt
kafka-python-ng==2.2.3
minio==7.2.15